adding support to compute all seeds with one action, also adding ability to read blocks from tiledb or json when generating seeds

+++ blockgeneratormodel/model.go
@@ -13,9 +13,9 @@ type BlockGenerator struct {
 	Seed                 int64
 }
 type Block struct {
-	RealizationIndex int   `json:"realization_index" eventstore:"realization_index"`
-	BlockIndex       int   `json:"block_index" eventstore:"block_index"`
-	BlockEventCount  int   `json:"block_event_count" eventstore:"block_event_count"`
+	RealizationIndex int32 `json:"realization_index" eventstore:"realization_index"`
+	BlockIndex       int32 `json:"block_index" eventstore:"block_index"`
+	BlockEventCount  int32 `json:"block_event_count" eventstore:"block_event_count"`
 	BlockEventStart  int64 `json:"block_event_start" eventstore:"block_event_start"` //inclusive - will be one greater than previous event end
 	BlockEventEnd    int64 `json:"block_event_end" eventstore:"block_event_end"`     //inclusive - will be one less than event start if event count is 0.
 }
@@ -40,7 +40,7 @@ func (bg BlockGenerator) GenerateBlocks() []Block {
 	for {
 		events := int(poisson.Rand())
 		EventEnd = EventStart + (int64(events) - 1)
-		block := Block{BlockIndex: Index, RealizationIndex: Realization, BlockEventCount: events, BlockEventStart: EventStart, BlockEventEnd: EventEnd}
+		block := Block{BlockIndex: int32(Index), RealizationIndex: int32(Realization), BlockEventCount: int32(events), BlockEventStart: EventStart, BlockEventEnd: EventEnd}
 		blocks = append(blocks, block)
 		if Index == bg.BlocksPerRealization {
 			Realization++
